+++ packages/yourloops/models/enums/routes.enum.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023, Diabeloop
+ * Copyright (c) 2023-2024, Diabeloop
  *
  * All rights reserved.
  *
@@ -51,6 +51,7 @@ export enum AppRoute {
   NewConsent = '/new-consent',
   ProductLabelling = '/product-labelling',
   RenewConsent = '/renew-consent',
+  SignupInformation = '/signup-information',
   Training = '/training',
   VerifyEmail = '/verify-email',
   VerifyEmailResult = '/verify-email-result'