+++ test/locales/translations.test.js
@@ -80,7 +80,8 @@ const ignoredTransKeyInBlipFiles = [
   'params|${row.name}',
   '${titleType} Above Range',
   '${titleType} Below Range',
-  '${titleType} In Range'
+  '${titleType} In Range',
+  'Micro Bolus'
 ]
 const ignoredTransKeyForYourLoops = [
   // Countries (from locales/languages.json)
@@ -164,9 +165,11 @@ const ignoredTransKeyForYourLoops = [
   'alarm-dana-incompatible-actions-on-pump-title',
   'alarm-dana-occlusion-description',
   'alarm-hyperglycemia-description',
-  'alarm-hyperglycemia-title',
+  'alarm-hyperglycemia-title-old',
+  'alarm-hyperglycemia-title-new',
   'alarm-hypoglycemia-description',
-  'alarm-hypoglycemia-title',
+  'alarm-hypoglycemia-title-old',
+  'alarm-hypoglycemia-title-new',
   'alarm-insight-empty-insulin-cartridge-description',
   'alarm-insight-empty-insulin-cartridge-title',
   'alarm-insight-empty-pump-battery-description',
@@ -185,20 +188,23 @@ const ignoredTransKeyForYourLoops = [
   'alarm-kaleido-empty-pump-battery-title',
   'alarm-kaleido-insulin-cartridge-expired-description',
   'alarm-kaleido-occlusion-description',
-  'alarm-long-hyperglycemia-description',
+  'alarm-long-hyperglycemia-description-old',
+  'alarm-long-hyperglycemia-description-new',
   'alarm-long-hypoglycemia-description',
   'alarm-loop-mode-activated-description',
   'alarm-loop-mode-deactivated-description',
   'alarm-multiple-occurrences',
   'alarm-no-readings-hypoglycemia-risk-description',
-  'alarm-no-readings-hypoglycemia-risk-title',
+  'alarm-no-readings-hypoglycemia-risk-title-old',
+  'alarm-no-readings-hypoglycemia-risk-title-new',
   'alarm-occlusion-title',
   'alarm-pump-cannot-deliver-insulin-description',
   'alarm-sensor-session-expired-description-line1',
   'alarm-sensor-session-expired-description-line2',
   'alarm-sensor-session-expired-title',
   'alarm-sudden-rise-glycemia-description',
-  'alarm-sudden-rise-glycemia-title',
+  'alarm-sudden-rise-glycemia-title-old',
+  'alarm-sudden-rise-glycemia-title-new',
   'alarm-urgent-low-soon-description',
   'alarm-urgent-low-soon-title',
   'alarm-with-code',
@@ -206,6 +212,8 @@ const ignoredTransKeyForYourLoops = [
   'alert-invite-patient-failed-already-invited',
   'alert-reinvite-patient-failure',
   'alert-reinvite-patient-from-team-success',
+  'Basal',
+  'Bolus',
   'current-trigger-setting-hypoglycemia',
   'current-trigger-setting-tir',
   'product-labelling',
@@ -215,6 +223,7 @@ const ignoredTransKeyForYourLoops = [
   'optional',
   'training-body',
   'refresh',
+  'insulin-unit-u',
   'login-page-desktop-info-1',
   'login-page-desktop-info-2',
   'login-page-desktop-info-3',
@@ -286,7 +295,8 @@ const ignoredTransKeyForYourLoops = [
   'verify-email-error-title',
   'tooltip-total-derived-carbs',
   'sensor-warmup',
-  'sensor-warmup-session-end'
+  'sensor-warmup-session-end',
+  'signup-information-message-2'
 ]
 const ignoredTransKeyInYourLoopsFiles = [
   'yourloops|${s}',
@@ -299,7 +309,6 @@ const ignoredTransKeyInYourLoopsFiles = [
   'IMEI',
   'Identifier',
   'Name',
-  'Parameters',
   'params|${parameter.name}',
   'Device',
   'Unit',
@@ -315,6 +324,7 @@ const ignoredTransKeyInYourLoopsFiles = [
   'Pump version',
   'Serial Number',
   'Setting',
+  'Settings on day',
   'Time In Range'
 ]
 
@@ -445,7 +455,6 @@ describe('Locales tests', () => {
         missingTranslations.push(key)
       }
     }
-
     expect(unusedTranslations, `Unused translations: ${JSON.stringify(unusedTranslations)}`).to.be.empty
     expect(missingTranslations, `Missing translations: ${JSON.stringify(missingTranslations)}`).to.be.empty
   })